Small additions
I added a controller for the header, and started building it out.
Modified the header directive to use the new controller. Added a
resources directory under app/ and created a config file to build out
that module. Added a userState service/singleton to handle communication
with the server to determine the state of the the user (logged in,
active, things of that nature). Brought over the old out-dated favicon
from lostrequiem.com.